:art: #1107 Refactoring around persistent resource services
* Pulled PersistenceSpaceIdentifier into PersistableResourceService
* Refactored PersistableResourceService.create to return a
PersistenceSpaceIdentifier
* Adapt code following these changes
